Fix verification of N, T and R phases
'CountPhases' method modified: receives the string wiring as parameter and returns a dictionary with 'phase' (F, N, T, R) and quantity pair.

Then, in the main the program iterates with each key of the dict returned from CountPhases.
With the F phase has a different actions: if there is only 1 'F' in the string wiring, this counts to "Phase A", if there are 2, counts to "Phase A" and "Phase B", until the "Phase C".
With the others: "N" -> "Neutro", "T" -> "Terra", "R" -> Retorno.